backend/pyproject.toml
[project]
name = "chewy-space"
version = "0.1.0"
description = "BBTalk Backend Service"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "Django>=5.2",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=4.3.0",
    "django-filter>=24.0",
    "drf-spectacular>=0.27.0",
    "gunicorn>=21.2.0",
    "pyjwt>=2.8.0",
    "cryptography>=41.0.0",
    "qiniu>=7.12.0",
    "requests>=2.31.0",
    "psycopg2-binary>=2.9.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
